+++ src/Products/PloneMeeting/browser/overrides.py
@@ -210,7 +210,6 @@ class PMConfigActionsPanelViewlet(ActionsPanelViewlet):
     """Render actionspanel viewlet differently for elements of the MeetingConfig."""
 
     backPages = {'categories': 'data',
-                 'classifiers': 'data',
                  'itemtemplates': 'data',
                  'podtemplates': 'doc',
                  'recurringitems': 'data', }