+++ src/Products/PloneMeeting/setuphandlers.py
@@ -51,7 +51,6 @@ indexInfos = {
     'getCategory': ('FieldIndex', {}),
     'getItemIsSigned': ('FieldIndex', {}),
     'getItemNumber': ('FieldIndex', {}),
-    'getRawClassifier': ('FieldIndex', {}),
     'getProposingGroup': ('FieldIndex', {}),
     'getGroupsInCharge': ('KeywordIndex', {}),
     'getAssociatedGroups': ('KeywordIndex', {}),
@@ -85,8 +84,8 @@ columnInfos = ('getDate',
                'getPreferredMeeting', 'getPreferredMeetingDate',
                'linkedMeetingDate', 'linkedMeetingUID',
                'title_or_id', 'toDiscuss',
-               'privacy', 'pollType', 'listType', 'getItemNumber',
-               'getCategory', 'getRawClassifier')
+               'privacy', 'pollType', 'listType',
+               'getItemNumber', 'getCategory')
 transformsToDisable = ['word_to_html', 'pdf_to_html', 'pdf_to_text']