+++ src/Products/PloneMeeting/tests/testMeetingItem.py
@@ -101,8 +101,8 @@ class testMeetingItem(PloneMeetingTestCase):
     '''Tests the MeetingItem class methods.'''
 
     def test_pm_SelectableCategories(self):
-        '''Categories are available if isSelectable returns True.  By default,
-           isSelectable will return active categories for wich intersection
+        '''Categories are available if is_selectable returns True.  By default,
+           is_selectable will return active categories for wich intersection
            between meetingcategory.using_groups and current member
            proposingGroups is not empty.'''
         # Use MeetingCategory as categories
@@ -113,35 +113,26 @@ class testMeetingItem(PloneMeetingTestCase):
         # create an item for test
         self.changeUser('pmCreator1')
         expectedCategories = ['deployment', 'maintenance', 'development', 'events', 'research', 'projects', ]
-        expectedClassifiers = ['classifier1', 'classifier2', 'classifier3', ]
         # By default, every categories are selectable
         self.assertEqual([cat.id for cat in cfg.getCategories()], expectedCategories)
-        # And the behaviour is the same for classifiers
-        self.assertEqual([cat.id for cat in cfg.getCategories(classifiers=True)], expectedClassifiers)
-        # Deactivate a category and a classifier
+        # Deactivate a category
         self.changeUser('admin')
-        self.wfTool.doActionFor(cfg.categories.deployment, 'deactivate')
-        self.wfTool.doActionFor(cfg.classifiers.classifier2, 'deactivate')
+        self._disableObj(cfg.categories.deployment)
         expectedCategories.remove('deployment')
-        expectedClassifiers.remove('classifier2')
         # getCategories has caching in the REQUEST, we need to wipe this out
         self.cleanMemoize()
         self.changeUser('pmCreator1')
         # A deactivated category will not be returned by getCategories no matter an item is given or not
         self.assertEqual([cat.id for cat in cfg.getCategories()], expectedCategories)
-        self.assertEqual([cat.id for cat in cfg.getCategories(classifiers=True)], expectedClassifiers)
         # Specify that a category is restricted to some groups pmCreator1 is not creator for
         self.changeUser('admin')
         cfg.categories.maintenance.using_groups = (self.vendors_uid,)
-        cfg.classifiers.classifier1.using_groups = (self.vendors_uid,)
         expectedCategories.remove('maintenance')
-        expectedClassifiers.remove('classifier1')
         # getCategories has caching in the REQUEST, we need to wipe this out
         self.cleanMemoize()
         self.changeUser('pmCreator1')
         # if current user is not creator for one of the using_groups defined for the category, he can not use it
         self.assertEqual([cat.id for cat in cfg.getCategories()], expectedCategories)
-        self.assertEqual([cat.id for cat in cfg.getCategories(classifiers=True)], expectedClassifiers)
         # cfg.getCategories can receive a userId
         # pmCreator2 has an extra category called subproducts
         expectedCategories.append('subproducts')
@@ -441,7 +432,7 @@ class testMeetingItem(PloneMeetingTestCase):
         # the original item is no more sendable to the same meetingConfig
         self.failIf(item.mayCloneToOtherMeetingConfig(otherMeetingConfigId))
         # while cloning to another meetingConfig, some fields that are normally kept
-        # while duplicating an item are no more kept, like category or classifier that
+        # while duplicating an item are no more kept, like category that
         # depends on the meetingConfig the item is in
         self.assertNotEqual(newItem.getCategory(), item.getCategory())
         # if we remove the newItem, the reference in the original item annotation is removed
@@ -778,7 +769,7 @@ class testMeetingItem(PloneMeetingTestCase):
         self.changeUser('admin')
         for at in (cfg.annexes_types.item_annexes.objectValues() +
                    cfg.annexes_types.item_decision_annexes.objectValues()):
-            at.enabled = False
+            self._disableObj(at)
         # no available annex types, try to clone newItem now
         self.changeUser('pmManager')
         clonedItem = originalItem.clone(copyAnnexes=True, copyDecisionAnnexes=True)
@@ -1713,7 +1704,7 @@ class testMeetingItem(PloneMeetingTestCase):
 
         # disable category
         self.changeUser('siteadmin')
-        self.do(category, 'deactivate')
+        self._disableObj(category)
         self.assertFalse(category in cfg.getCategories(onlySelectable=True))
         self.changeUser('pmCreator1')
         new_item_without_category = item.clone()
@@ -3496,7 +3487,7 @@ class testMeetingItem(PloneMeetingTestCase):
         item2._update_after_edit()
         # a disabled category will still be displayed in the vocab if it is the currently used value
         self.changeUser('siteadmin')
-        self.do(cfg.categories.development, 'deactivate')
+        self._disableObj(cfg.categories.development)
         self.assertEqual(item.listCategories().values(),
                          [u'Development topics', u'Events', u'Research topics'])
         self.assertEqual(item2.listCategories().values(),
@@ -6585,7 +6576,7 @@ class testMeetingItem(PloneMeetingTestCase):
         self.assertEqual(item4.getItemReference(), 'Ref. 20170303/1')
 
     def test_pm_ItemReferenceUpdateWhenSpecificItemFieldsModified(self):
-        """When a item is modified, if 'category', 'classifier', 'proposingGroup'
+        """When a item is modified, if 'category', 'proposingGroup'
            or 'otherMeetingConfigsClonableTo' field is changed, we need to update
            every itemReference starting from current item."""
         self.changeUser('siteadmin')
@@ -6601,7 +6592,6 @@ class testMeetingItem(PloneMeetingTestCase):
             "python: here.getMeeting().getDate().strftime('%Y%m%d') + '/' + "
             "str(here.getProposingGroup(True).get_acronym()) + '/' + "
             "str(here.getCategory()) + '/' + "
-            "str(here.getClassifier() and here.getClassifier().getId() or '-') + '/' + "
             "('/'.join(here.getOtherMeetingConfigsClonableTo()) or '-') + '/' + "
             "here.Title() + '/' + "
             "str(here.getItemNumber(relativeTo='meetingConfig', for_display=True))")
@@ -6610,29 +6600,25 @@ class testMeetingItem(PloneMeetingTestCase):
         meeting = self.create('Meeting', date=DateTime('2017/03/03'))
         self.presentItem(item)
         self.freezeMeeting(meeting)
-        self.assertEqual(item.getItemReference(), '20170303/Devel/development/-/-/Title1/1')
+        self.assertEqual(item.getItemReference(), '20170303/Devel/development/-/Title1/1')
         # change category
         item.setCategory('research')
         item._update_after_edit()
-        self.assertEqual(item.getItemReference(), '20170303/Devel/research/-/-/Title1/1')
-        # change classifier
-        item.setClassifier(cfg.classifiers.classifier1.UID())
-        item._update_after_edit()
-        self.assertEqual(item.getItemReference(), '20170303/Devel/research/classifier1/-/Title1/1')
+        self.assertEqual(item.getItemReference(), '20170303/Devel/research/-/Title1/1')
         # change proposingGroup
         item.setProposingGroup(self.vendors_uid)
         item._update_after_edit()
-        self.assertEqual(item.getItemReference(), '20170303/Devil/research/classifier1/-/Title1/1')
+        self.assertEqual(item.getItemReference(), '20170303/Devil/research/-/Title1/1')
         # change otherMeetingConfigsClonableTo
         item.setOtherMeetingConfigsClonableTo((cfg2Id,))
         item._update_after_edit()
         self.assertEqual(item.getItemReference(),
-                         '20170303/Devil/research/classifier1/{0}/Title1/1'.format(cfg2Id))
+                         '20170303/Devil/research/{0}/Title1/1'.format(cfg2Id))
         # changing the Title will not update the reference
         item.setTitle('Title2')
         item._update_after_edit()
         self.assertEqual(item.getItemReference(),
-                         '20170303/Devil/research/classifier1/{0}/Title1/1'.format(cfg2Id))
+                         '20170303/Devil/research/{0}/Title1/1'.format(cfg2Id))
 
     def test_pm_ItemReferenceUpdateWhenItemPositionChangedOnMeeting(self):
         """When an item position changed in the meeting, the itemReference is updated."""